file

+++ TeamProject_HONJAL/Honjal/src/main/java/com/honjal/honjal/dao/ext/FileDao.java
@@ -5,13 +5,15 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 import com.honjal.honjal.dao.GenericDao;
+import com.honjal.honjal.model.ContentFilesDTO;
 import com.honjal.honjal.model.FileDTO;
 
-public interface FileDao extends GenericDao<FileDao, Integer> {
-	
+public interface FileDao extends GenericDao<FileDao, Long> {
+
 	public int insertOrUpdate(FileDTO fileDto);
 	public int insertWithList(@Param("filesList") List<FileDTO>filesList);
 	public int insertOrUpdateWithList(@Param("filesList") List<FileDTO>filesList);
+
 
 
 }
